File system functions moved to fs.hpp
