+++ include/utils.hpp
@@ -11,11 +11,3 @@ std::string rtrim(const std::string& s);
 // IO () -> IO (Maybe String)
 std::optional<std::string> fzf();
 std::optional<std::string> fzf_dir();
-// String -> IO Bool
-bool set_working_dir(const std::string& path);
-
-// String -> IO Bool
-bool create_dir(const std::string& path);
-
-bool move_file(const std::string& src, const std::string& dest);
-bool copy_file(const std::string& src, const std::string& dest);
