+++ src/utils.cpp
@@ -72,51 +72,3 @@ std::optional<std::string> fzf_dir() {
     return output;
 }
 
-bool set_working_dir(const std::string& path) {
-    std::string expanded_path = expand_tilde(path);
-    const auto base_path = std::filesystem::path{expanded_path};
-    try {
-        std::filesystem::current_path(base_path);
-    }
-    catch (std::filesystem::filesystem_error& error) {
-        std::cerr << "Error changing directory: " << error.what() << std::endl;
-        return false;
-    }
-
-    return true;
-}
-
-bool create_dir(const std::string& path) {
-    try {
-        const auto dir_path = std::filesystem::path{expand_tilde(path)};
-
-        std::filesystem::create_directory(dir_path);
-    }
-    catch (std::filesystem::filesystem_error& error) {
-        std::cerr << "Error creating dir: " << error.what() << std::endl;
-        return false;
-    }
-    return true;
-}
-
-bool move_file(const std::string& src, const std::string& dest) {
-    try {
-        std::filesystem::rename(src, dest);
-    }
-    catch (std::filesystem::filesystem_error &e) {
-        std::cerr << "Error moving file: " << e.what() << std::endl;
-        return false;
-    }
-    return true;
-}
-
-bool copy_file(const std::string& src, const std::string& dest) {
-    try {
-        std::filesystem::copy(src, dest);
-    }
-    catch (std::filesystem::filesystem_error &e) {
-        std::cerr << "Error copying file: " << e.what() << std::endl;
-        return false;
-    }
-    return true;
-}
